feat: add linglong trigger
add linglong trigger

Log: add linglong trigger
