+++ cmd/root.go
@@ -1,0 +1,32 @@
+/*
+Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+
+*/
+package cmd
+
+import (
+	"log"
+
+	"github.com/spf13/cobra"
+)
+
+// rootCmd represents the base command when called without any subcommands
+var rootCmd = &cobra.Command{
+	Use:   "jenkins-bridge-client",
+	Short: "A client to call jenkins-bridge api",
+	Long: `A client to call jenkins-bridge api.
+you can trigger build, api-check,linglong build and so on.`,
+}
+
+// Execute adds all child commands to the root command and sets flags appropriately.
+// This is called by main.main(). It only needs to happen once to the rootCmd.
+func Execute() {
+	err := rootCmd.Execute()
+	if err != nil {
+		log.Fatal(err.Error())
+	}
+}
+
+func init() {
+
+}
