+++ go.mod
@@ -5,12 +5,15 @@ go 1.17
 require (
 	github.com/dlclark/regexp2 v1.4.0
 	github.com/go-resty/resty/v2 v2.7.0
-	github.com/google/go-github/v44 v44.0.0
+	github.com/google/go-github v17.0.0+incompatible
 	github.com/myml/ghtoken v0.0.0-20220506023416-6750199c71f6
+	github.com/spf13/cobra v1.5.0
 )
 
 require (
+	github.com/google/go-cmp v0.5.7 // indirect
 	github.com/google/go-querystring v1.1.0 // indirect
-	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/net v0.0.0-20211029224645-99673261e6eb // indirect
 )
